--- a/src/content-scripts/ribbon.ts
+++ b/src/content-scripts/ribbon.ts
@@ -1,5 +1,0 @@
-import 'babel-polyfill'
-
-export async function main() {}
-
-main()