--- a/src/content-scripts/search-injection.ts
+++ b/src/content-scripts/search-injection.ts
@@ -1,0 +1,8 @@
+import 'babel-polyfill'
+import { initSearchInjection } from '../search-injection/content_script'
+
+export async function main() {
+    initSearchInjection()
+}
+
+main()