--- a/src/content-scripts/tooltip.ts
+++ b/src/content-scripts/tooltip.ts
@@ -1,0 +1,5 @@
+import 'babel-polyfill'
+
+export async function main() {}
+
+main()