--- a/src/content-scripts/types.ts
+++ b/src/content-scripts/types.ts
@@ -1,0 +1,4 @@
+export type ContentScript = 'search_injection' | 'in_page_ui'
+export interface ContentScriptRegistry {
+    registerInPageUIScript(main: () => Promise<void>): Promise<void>
+}