--- a/src/overview/sidebar-left/components/collections-container.tsx
+++ b/src/overview/sidebar-left/components/collections-container.tsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 
 import CollectionsButton from './collections-button'
-import { actions as acts } from 'src/overview/sidebar-left/'
+import * as acts from 'src/overview/sidebar-left//actions'
 import { selectors as filters, actions as filterActs } from 'src/search-filters'
 import { selectors as lists } from 'src/custom-lists'
 import * as selectors from '../selectors'
@@ -17,7 +17,4 @@ const mapDispatch = dispatch => ({
     onShowBtnClick: () => dispatch(filterActs.delListFilter()),
 })
 
-export default connect(
-    mapState,
-    mapDispatch,
-)(CollectionsButton)
+export default connect(mapState, mapDispatch)(CollectionsButton)