… package.json
… src/manifest.json
… src/overview/delete-confirm-modal/actions.ts
… src/overview/results/actions.ts
… src/overview/results/components/ResultListContainer.tsx
… src/search-filters/components/tags-filter-container.tsx
… src/search-filters/components/users-filter-container.tsx
… src/search-injection/components/container.tsx
… src/search-injection/constants.ts
… src/search-injection/detection.ts
… src/search-injection/dom.tsx
… src/search-injection/types.ts
… src/sidebar-overlay/actions.ts
… src/sidebar-overlay/annotation-box/annotation-box-container.tsx
… src/sidebar-overlay/comment-box/components/comment-box-container.tsx
… src/sidebar-overlay/comment-box/components/comment-box-form.tsx
… src/sidebar-overlay/comment-box/components/tag-input-container.tsx
… src/sidebar-overlay/content_script/rendering.ts
… src/sidebar-overlay/epics.ts
… src/sidebar-overlay/reducer.ts
… src/sidebar-overlay/ribbon-sidebar-controller/ribbon-sidebar-container.tsx
… src/sidebar-overlay/ribbon/components/ribbon-container.tsx
… src/sidebar-overlay/ribbon/components/ribbon.tsx
… src/sidebar-overlay/sidebar/components/filters-sidebar-container.tsx
… src/sidebar-overlay/sidebar/components/lists-container.js
… src/sidebar-overlay/sidebar/components/result-list-container.tsx
… src/sidebar-overlay/sidebar/components/sidebar-container.tsx
… src/sidebar-overlay/sidebar/reducer.ts
… src/social-integration/actions.ts
… src/social-integration/twitter/components/save-to-memex-container.tsx