In a search to reduce content script bloat, I made many imports more specific to prevent useless stuff from being imported included in the import script. The rabbit hole turned out to be too deep however, but in the search I did uncover `src/util/nlp-time-filter` and `src/highlighting/ui/anchoring/index.js` to be adding non-trivial amounts of bloat.