@blackforestboi 2020-03-03
It seems to only reduce the content `content_script.js.map` file, not the content script itself though, when I build it.