@ShishKabab 2020-03-03
Yup, in the end, size kept decreasing, then going back to normal again. Also the reduction at each cycle of diving kept converging towards the initial size. The most useful part of this exercise was discovering two big-ish libraries, but no immediate script reductions are gained from this work  :(