@ShishKabab 2020-03-25
I've started a crude prototype of the separate loading of the in page UI content script (in the end, I had to combine the sidebar, ribbon and content tooltip.) Main content script size is now 3,6MB unminified, UI script is 14MB. Initial test shows that using the development build, compiling time of content script goes down from >340ms to 56ms and total compile + execution time down from >500ms to 108ms. This is seems like a good improvement already.

The downsides:
* So far the UI script runs, but doesn't seem to work yet. Something seems to go wrong setting up the RPC methods, which should maybe work differently anyway. Also, I don't actually see any sidebar yet.  Some investigation needed there, may influence final numbers of perf gains.
* Bigger problem: Loading the sidebar on demand seems to take 880ms on my fast machine. That number'll probably go up when everything works. Breaking this number down further, the time it takes from asking the browser to inject the script, to executing the first line of the script before the imports already takes 686 ms, probably because of the big file size. From there, executing the imports takes another 134ms, then the actual execution (which doesn't do what it's supposed yet) takes 7ms. Minification might reduce some of that first 686ms.

Do we want to go further on this path? I don't know how many people have the sidebar always enabled, but having them not load when you've disabled both the sidebar and tooltip seems like a good thing to me. However, the question is whether we can get to a reasonable UX.