@ShishKabab 2020-03-25
Clarification: The content tooltip, ribbon and sidebar are comibined because they share state between them, and would need more work to be split out.