@ShishKabab 2020-03-25
I'd still like to know why we accumulate so much code in the UI script. Maybe it needs some more in-depth investigation on the logic.

By the way, I just ran a test using a production build. Main content script is 392KB now, UI script is 2,.2MB. Loading time is reduced from 686ms to 240ms.

This makes things more worthwile to explore. In any case, I'm getting the feeling it might be really worthwhile to split the ribbon/tooltip from the sidebar. The sidebar contains way more actual logic.