@ShishKabab 2020-03-25
So, I think I have the cause at least why the remote functions are not registered: script-level globals. The RPC file has a constant with all registered functions. The file gets compiled into both the main content script and UI content script, so they both have different globals. Going to create a hack to see if this is right.