@ShishKabab 2020-03-25
Yup, seems to work  :)  Continuing testing now to see if this can functionally completely work. If it does, we can look at splitting out more stuff.