@blackforestboi 2020-03-25
Great improvements, wow!

> This makes things more worthwile to explore. In any case, I'm getting the feeling it might be really worthwhile to split the ribbon/tooltip from the sidebar. The sidebar contains way more actual logic.

Yeah I agree, as noted in the slack thread: 
"My initial thought was that we could split out the ribbon & the tooltip from the sidebar and make them pure HTML, so they are super fast to load and open. When that happens, we could already hook the sidebar in the background, and profit from a delay in user interaction to reach the button to open the sidebar fully. That means we might not need to optimise the UI script size (of the sidebar) initially. " - but should optimise it definitely along the way.