@ShishKabab 2020-03-25
I think we don't even need to ditch React to start. As a first step, I'd just decouple both scripts and see what happens. I have some idea on how to do that without a big refactor, just to see what kind of impact it has.