@blackforestboi 2020-03-25
Ah ok, yeah right its super small. In that case a decoupling of those parts and loading of the sidebar when the ribbon is activated should already bring some good improvements.