@ShishKabab 2020-03-26
OK, I was too positive on the effort required to separate the sidebar and the ribbon, even with the hack I had in mind. There's just a lot of coupling. Aside from them sharing a Redux state, and the logic being tangled up between the sidebar, ribbon and the thing controlling both of them, they also share a React context. Picking this apart will take a lot of work, no telling how deep the rabbit hole goes. 3 full time days minimum I'd, plus other stuff that will come up. But, if we want to go to a light-weight version of the ribbon, I'd say this is the only way to go.

This may be a good time to re-evaluate how we're tackling this.