@blackforestboi 2020-04-17
closed in favor of https://github.com/WorldBrain/Memex/issues/980